Move statistics to feature folder
