+++ server/src/features/statistics/dataFixer.ts
