+++ server/src/features/statistics/datafiles/ropecon/2017/games.json
