+++ server/src/features/statistics/datafiles/ropecon/2018/games.json
