+++ server/src/features/statistics/fixer-helpers/wildFix.ts
@@ -1,4 +1,4 @@
-import { readJson, writeJson } from 'server/statistics/statsUtil';
+import { readJson, writeJson } from 'server/features/statistics/statsUtil';
 import { logger } from 'server/utils/logger';
 
 export const wildFix = (year: number, event: string, type: string): void => {
