+++ server/src/features/statistics/statGenerator.ts
