+++ server/src/features/statistics/statistics-helpers/getResultsStats.ts
