+++ server/src/features/statistics/statistics-helpers/resultDataHelpers.ts
@@ -2,7 +2,7 @@ import { logger } from 'server/utils/logger';
 import { Game } from 'shared/typings/models/game';
 import { ResultsCollectionEntry } from 'server/typings/result.typings';
 import { StringNumberObject } from 'server/typings/common.typings';
-import { toPercent } from 'server/statistics/statsUtil';
+import { toPercent } from 'server/features/statistics/statsUtil';
 
 export const getSignupsByTime = (
   results: readonly ResultsCollectionEntry[]
