+++ server/src/features/statistics/statsUtil.ts
